conventions module
