+++ core/effects/importmap.ts
@@ -1,5 +1,5 @@
 import { join } from "@std/path";
-import { generatedFolder } from "../constants.ts";
+import { generatedFolder } from "$lib/conventions.ts";
 import { createEffect, type Effect } from "@radish/effect-system";
 
 export interface ImportMap {
