+++ core/mod.ts
@@ -1,6 +1,6 @@
 export { importmapPath } from "$effects/importmap.ts";
 export { manifestPath } from "$effects/manifest.ts";
 export { onDispose } from "./cleanup.ts";
-export { globals } from "./constants.ts";
+export * from "./conventions.ts";
 export { startApp } from "./start.ts";
 export type { Config, ManifestBase } from "./types.d.ts";
