+++ core/plugins/router/router.ts
@@ -3,7 +3,7 @@ import { manifest } from "$effects/manifest.ts";
 import { build } from "$effects/mod.ts";
 import type { Manifest } from "$effects/render.ts";
 import { type Route, type RouteContext, router } from "$effects/router.ts";
-import { routesFolder } from "$lib/constants.ts";
+import { routesFolder } from "$lib/conventions.ts";
 import { manifestShape } from "$lib/plugins/render/hooks/manifest.ts";
 import type { MaybePromise } from "$lib/types.d.ts";
 import { createStandardResponse } from "$lib/utils/http.ts";
